+++ crates/micros/src/mcu/_.rs
@@ -1,6 +1,6 @@
 //
 #![doc = crate::_DOC_MCU!()] // public
-#![doc = crate::_doc!(modules: crate; mcu: avr, esp32)]
+#![doc = crate::_doc!(modules: crate; mcu: avr, esp32, sam)]
 #![doc = crate::_doc!(flat:"mcu")]
 #![doc = crate::_doc!(hr)]
 //!
@@ -15,10 +15,11 @@
 //!
 //! # Microcontrollers
 //!
-//! | MCU             | Core              | Memory                                 | GPIO | Timers              | Serial                      | Analog        | Radio        |
-//! | --------------- | ----------------- | -------------------------------------- | ---: | ------------------- | --------------------------- | ------------- | ------------ |
-//! | `McuAtmega328p` | AVR 8-bit, 20 MHz | 32 KiB Flash, 2 KiB SRAM, 1 KiB EEPROM |   23 | T0/T1/T2            | USART0, ◇SPI, ◇TWI          | ◇10-bit ADC   | —            |
-//! | `McuEsp32C3`    | RV32IMC, 160 MHz  | 384 KiB ROM, 400 KiB SRAM, ext. Flash  | ≤ 22 | ◇GPTimer, ◇SYSTIMER | UART0, I²C0, ◇UART1, ◇SPI   | ◇2x12-bit ADC | ◇Wi-Fi, ◇BLE |
+//! | MCU             | Core              | Memory                                 | GPIO | Timers              | Serial                       | Analog          | Radio        |
+//! | --------------- | ----------------- | -------------------------------------- | ---: | ------------------- | ---------------------------- | --------------- | ------------ |
+//! | `McuAtmega328p` | AVR 8-bit, 20 MHz | 32 KiB Flash, 2 KiB SRAM, 1 KiB EEPROM |   23 | T0/T1/T2            | USART0, ◇SPI, ◇TWI           | ◇10-bit ADC     | —            |
+//! | `McuEsp32C3`    | RV32IMC, 160 MHz  | 384 KiB ROM, 400 KiB SRAM, ext. Flash  | ≤ 22 | ◇GPTimer, ◇SYSTIMER | UART0, I²C0, ◇UART1, ◇SPI    | ◇2x12-bit ADC   | ◇Wi-Fi, ◇BLE |
+//! | `McuSam3x8e`    | Cortex-M3, 84 MHz | 512 KiB Flash, 96 KiB SRAM             |  103 | ◇TC0/TC1/TC2        | ◇UART, ◇USART0–3, ◇SPI, ◇TWI | ◇12-bit ADC/DAC | —            |
 //!
 //! ```txt
 //! ◇  hardware capability not yet exposed by this crate
@@ -32,6 +33,8 @@ crate::mods_in! {
     pub mod_ avr;
     #[cfg_attr(not(nightly_doc), cfg(feature = "esp32"))]
     pub mod_ esp32;
+    #[cfg_attr(not(nightly_doc), cfg(feature = "sam"))]
+    pub mod_ sam;
 }
 crate::mods_out! { // _pub_mods, _reexports
     _pub_mods {
@@ -39,6 +42,8 @@ crate::mods_out! { // _pub_mods, _reexports
         pub use super::avr::_all::*;
         #[cfg_attr(not(nightly_doc), cfg(feature = "esp32"))]
         pub use super::esp32::_all::*;
+        #[cfg_attr(not(nightly_doc), cfg(feature = "sam"))]
+        pub use super::sam::_all::*;
     }
     _reexports {
         #[doc(inline)]
@@ -47,5 +52,8 @@ crate::mods_out! { // _pub_mods, _reexports
         #[doc(inline)]
         #[cfg(feature = "esp32c3")]
         pub use super::esp32::McuEsp32C3;
+        #[doc(inline)]
+        #[cfg(feature = "sam3x8e")]
+        pub use super::sam::McuSam3x8e;
     }
 }